Feature/remove seqjson and seqn lang specifics (#1446)
* Moved the lint calls outside of lang specific files

* Added a default adaptation and hooks for auto indent and auto complete

* Fixed filterText not having a type and reconfiguring the sequence editor

* Removed a lint problem

* Fixed an issue where reconfiguring the seqn editor was throwing away all extensions

* Moved auto complete to a comparment
